refactor: OAuth 連携部分を切り出し
